refactor: UserInfoService에서 Change 관련 UseCase 분리
